refactor: Rename folder name
